deploymentML1P2/backend/requirements.txt
waitress
pandas
numpy
seaborn
matplotlib
flask
scikit-learn==1.0.2
tensorflow-cpu